RBGL.overview {RBGL}	R Documentation

RBGL.overview

Description

The RBGL package consists of a number of interfaces to the Boost C++ library for graph algorithms. This page follows, approximately, the chapter structure of the monograph on the Boost Graph Library by Siek et al., and gives hyperlinks to documentation on R functions currently available, along with the names of formal parameters to these functions.

basicAlgs

Functions 	 parameters

bandwidth 	 g

bfs 	 object,node,checkConn

dfs 	 object,node,checkConn

edgeConnectivity 	 g

gprofile 	 g

isomorphism 	 g1,g2

minCut 	 g

transitive.closure 	 g

tsort 	 x

ShortestPaths

Functions 	 parameters

bellman.ford.sp 	 g,start

dag.sp 	 g,start

dijkstra.sp 	 g,start

extractPath 	 s,f,pens

johnson.all.pairs.sp 	 g

sp.between 	 g,start,finish

sp.between.old 	 g,start,finish

sp.between.scalar 	 g,start,finish

MinimumSpanningTree

Functions 	 parameters

mstree.kruskal 	 x

ConnectedComponents

Functions 	 parameters

connectedComp 	 g

highlyConnSG 	 g,sat,ldv

incremental.components 	 g

init.incremental.components 	 g

same.component 	 g,node1,node2

strongComp 	 g

MaximumFlow

Functions 	 parameters

edmonds.karp.max.flow 	 g,source,sink

push.relabel.max.flow 	 g,source,sink

SparseMatrixOrdering

Functions 	 parameters

cuthill.mckee.ordering 	 g

minDegreeOrdering 	 g,delta

sloan.ordering 	 g,w1,w2

LayoutAlgorithms

Functions 	 parameters

circle.layout 	 g,radius

kamada.kawai.spring.layout 	 g,edge_or_side,es_length

GraphClustering

Functions 	 parameters

betweenness.centrality.clustering 	 g,threshold,normalize

Betweenness

Functions 	 parameters

brandes.betweenness.centrality 	 g

Wavefront

Functions 	 parameters

aver.wavefront 	 g

ith.wavefront 	 g,start

maxWavefront 	 g

rms.wavefront 	 g

References

Boost Graph Library ( www.boost.org/libs/graph/doc/index.html )

The Boost Graph Library: User Guide and Reference Manual; by Jeremy G. Siek, Lie-Quan Lee, and Andrew Lumsdaine; (Addison-Wesley, Pearson Education Inc., 2002), xxiv+321pp. ISBN 0-201-72914-8

[Package RBGL version 1.58.0.1 Index]