betweenness.centrality.clustering {RBGL}	R Documentation

Graph clustering based on edge betweenness centrality

Description

Graph clustering based on edge betweenness centrality

Usage

betweenness.centrality.clustering(g, threshold = -1, normalize = TRUE)

Arguments

g
an instance of the graph class with edgemode “undirected”

threshold
threshold to terminate clustering process

normalize
boolean, when TRUE, the edge betweenness centrality is scaled by 2/((n-1)(n-2)) where n is the number of vertices in g; when FALSE, the edge betweenness centrality is the absolute value

Details

To implement graph clustering based on edge betweenness centrality.

The algorithm is iterative, at each step it computes the edge betweenness centrality and removes the edge with maximum betweenness centrality when it is above the given threshold. When the maximum betweenness centrality falls below the threshold, the algorithm terminates.

See documentation on Clustering algorithms in Boost Graph Library for details.

Value

A list of

no.of.edges
number of remaining edges after removal

edges
remaining edges

edge.betweenness.centrality
betweenness centrality of remaining edges

Author(s)

Li Long <li.long@isb-sib.ch>

References

Boost Graph Library ( www.boost.org/libs/graph/doc/index.html )

The Boost Graph Library: User Guide and Reference Manual; by Jeremy G. Siek, Lie-Quan Lee, and Andrew Lumsdaine; (Addison-Wesley, Pearson Education Inc., 2002), xxiv+321pp. ISBN 0-201-72914-8

See Also

brandes.betweenness.centrality

Examples

con <- file(system.file("XML/conn.gxl",package="RBGL"))
coex <- fromGXL(con)
close(con)
coex <- ugraph(coex)
betweenness.centrality.clustering(coex, 0.5, TRUE)

[Package RBGL version 1.58.0.1 Index]