graphGenerator {RBGL}	R Documentation

Generate an undirected graph with adjustable clustering coefficient

Description

Generate an undirected graph with adjustable clustering coefficient

Usage

graphGenerator(n, d, o)

Arguments

n
no. of nodes in the generated graph

d
parameter for preferential attachment

o
parameter for triple generation

Details

The graph generator works according to the prefential attachment rule. It also generates graphs with adjustable clustering coefficient. Parameter d specifies how many preferred edges a new node has. Parameter o limits how many triples to add to a new node.

See reference for details.

Value

no. of nodes
No. of nodes in the generated graph

no. of edges
No. of edges in the generated graph

edges
Edges in the generated graph

Author(s)

Li Long <li.long@isb-sib.ch>

References

Approximating Clustering Coefficient and Transitivity, T. Schank, D. Wagner, Journal of Graph Algorithms and Applications, Vol. 9, No. 2 (2005).

See Also

clusteringCoef, transitivity, clusteringCoefAppr

Examples

n <- 20
d <- 6
o <- 3
gg <- graphGenerator(n, d, o)

[Package RBGL version 1.58.0.1 Index]