lambdaSets {RBGL}	R Documentation

Find all the lambda-sets in an undirected graph

Description

Find all the lambda-sets in an undirected graph

Usage

lambdaSets(g)

Arguments

g
an instance of the graph class

Details

From reference (1), p. 270: A set of nodes is a lambda-set if any pair of nodes in the lambda set has larger edge connectivity than any pair of nodes consisting of one node from within the lamda set and a second node from outside the lamda set.

As stated in reference (2), a lambda set is a maximal subset of nodes who have more edge-independent paths connecting them to each other than to outsiders.

A lambda set could be characterized by the minimum edge connectivity k among its members, and could be called lambda-k sets.

Let N be maximum edge connectivity of graph g, we output all the lambda-k set for all k = 1, ..., N.

Value

Maximum edge connectivity, N, of the graph g, and A list of length N; k-th entry (k = 1, ..., N) is a list of all the lambda-k sets in graph g.

Author(s)

Li Long <li.long@isb-sib.ch>

References

(1) Social Network Analysis: Methods and Applications. By S. Wasserman and K. Faust, pp. 269. (2) LS sets, lambda sets and other cohesive subsets. By S. P. Borgatti, M. G. Everett, P. R. Shirey, Social Networks 12 (1990) p. 337-357

Examples

con <- file(system.file("XML/snalambdaex.gxl",package="RBGL"))
coex <- fromGXL(con)
close(con)

lambdaSets(coex)

[Package RBGL version 1.58.0.1 Index]