mstree.prim {RBGL}	R Documentation

Compute minimum spanning tree for an undirected graph

Description

Compute minimum spanning tree for an undirected graph

Usage

mstree.prim(g)

Arguments

g
an instance of the graph class with edgemode “undirected”

Details

This is Prim's algorithm for solving the minimum spanning tree problem for an undirected graph with weighted edges.

See documentations on this function in Boost Graph Library for more details.

Value

A list of

edges
the edges that form the minimum spanning tree

weights
the total weight of the minimum spanning tree

Author(s)

Li Long <li.long@isb-sib.ch>

References

Boost Graph Library ( www.boost.org/libs/graph/doc/index.html )

The Boost Graph Library: User Guide and Reference Manual; by Jeremy G. Siek, Lie-Quan Lee, and Andrew Lumsdaine; (Addison-Wesley, Pearson Education Inc., 2002), xxiv+321pp. ISBN 0-201-72914-8

See Also

mstree.kruskal

Examples

con <- file(system.file("XML/conn2.gxl",package="RBGL"))
coex <- fromGXL(con)
close(con)

mstree.prim(coex)

[Package RBGL version 1.58.0.1 Index]